+++ JavaScript/test/runner-cultures.js
@@ -1,0 +1,7 @@
+var Culture = require('../compiled/culture').Culture;
+
+// List of supported cultures
+module.exports = {
+    'Eng': Culture.English,
+    'Spa': Culture.Spanish
+};
